refactor: reorganize imports and clean up unused code in various modules moving tracked_any functions
