+++ src/v4vapp_backend_v2/actions/tracked_any.py
@@ -1,6 +1,6 @@
 from typing import Annotated, Any
 
-from pydantic import BaseModel, Discriminator, Tag
+from pydantic import BaseModel, Discriminator, Tag, ValidationError
 
 from v4vapp_backend_v2.actions.tracked_models import TrackedBaseModel
 from v4vapp_backend_v2.config.setup import InternalConfig
@@ -9,7 +9,7 @@ from v4vapp_backend_v2.hive_models.op_fill_order import FillOrder
 from v4vapp_backend_v2.hive_models.op_fill_recurrent_transfer import FillRecurrentTransfer
 from v4vapp_backend_v2.hive_models.op_limit_order_create import LimitOrderCreate
 from v4vapp_backend_v2.hive_models.op_recurrent_transfer import RecurrentTransfer
-from v4vapp_backend_v2.hive_models.op_transfer import Transfer
+from v4vapp_backend_v2.hive_models.op_transfer import Transfer, TransferBase
 from v4vapp_backend_v2.models.invoice_models import Invoice
 from v4vapp_backend_v2.models.payment_models import Payment
 
@@ -120,8 +120,7 @@ async def load_tracked_object(tracked_obj: TrackedAny | str) -> TrackedAny | Non
                     answer = DiscriminatedTracked.model_validate(value)
                     return answer.value
 
-    if getattr(tracked_obj, "collection_name", None):
-        collection_name = tracked_obj.collection_name
+    if collection_name := getattr(tracked_obj, "collection_name", None):
         result = await InternalConfig.db[collection_name].find_one(
             filter=tracked_obj.group_id_query,
         )
@@ -132,4 +131,65 @@ async def load_tracked_object(tracked_obj: TrackedAny | str) -> TrackedAny | Non
     return None
 
 
-# End of the file
+def tracked_any_filter(tracked: dict[str, Any]) -> TrackedAny:
+    """
+    Validates and filters a tracked object, ensuring it is of type OpAny, Invoice, or Payment.
+
+    Removes the '_id' field from the input dictionary if present, then attempts to validate
+    the object using the DiscriminatedTracked model. If validation is successful, returns
+    the validated object as a TrackedAny type. Raises a ValueError if validation fails.
+
+    Args:
+        tracked (dict[str, Any]): The tracked object to validate and filter.
+
+    Returns:
+        TrackedAny: The validated tracked object of type OpAny, Invoice, or Payment.
+
+    Raises:
+        ValueError: If the object cannot be validated as one of the expected types.
+
+    """
+    if "_id" in tracked:
+        del tracked["_id"]  # Remove _id field if present
+
+    try:
+        value = {"value": tracked}
+        answer = DiscriminatedTracked.model_validate(value)
+        return answer.value
+    except ValidationError as e:
+        raise ValueError(f"Failed to validate tracked object: {e}") from e
+    except ValueError as e:
+        logger.warning(
+            f"Parsing as OpAny, Invoice, or Payment. {e}",
+            extra={"notification": False, "tracked": tracked},
+        )
+        raise ValueError(
+            f"Invalid tracked object type: Expected OpAny, Invoice, or Payment. {e}"
+        ) from e
+
+
+def tracked_transfer_filter(tracked: dict[str, Any]) -> TrackedTransfer:
+    """
+    Validates and filters a tracked object, ensuring it is of type TrackedTransfer.
+
+    Removes the '_id' field from the input dictionary if present, then attempts to validate
+    the object using the TrackedTransfer model. If validation is successful, returns
+    the validated object as a TrackedTransfer type. Raises a ValueError if validation fails.
+
+    Args:
+        tracked (dict[str, Any]): The tracked object to validate and filter.
+
+    Returns:
+        TrackedTransfer: The validated tracked object of type TrackedTransfer.
+
+    Raises:
+        ValueError: If the object cannot be validated as a TrackedTransfer.
+    """
+    tracked_any = tracked_any_filter(tracked)
+    if isinstance(tracked_any, (TransferBase, Transfer, RecurrentTransfer, FillRecurrentTransfer)):
+        return tracked_any
+    raise ValueError(
+        f"Invalid tracked object type: Expected TrackedTransfer, got {type(tracked_any)}"
+    )
+
+
